Mass reindentation with gindent(1)
In order for b-em to remain consistent with more than one person working on
it, it's important to have consistency within the code.

The suggestion is to use GNU Indent (gindent(1)) to do this -- and the Linux
kernel style seems appropriate here, hence:

    for i in src/*.[ch]; do echo "$i" && gindent -linux -l78 -di16 "$i"; done

Eventually, this could be automated, but this seems like the easiest set of
rules to standardise on.

Note that this deliberately just targets the code under src/ and not anything
else -- as this is core b-em, and I don't want to touch other third-party code
we also choose to include as part of b-em.