general: make opening a file a fatal error in many fewer cases
There are many cases in which the emulator can reasonably continue
after an attempt to open a file has failed.  Many of these already had
code to handle them but much of this was broken by x_fopen which made
every attempt to open a file a fatal error.  This commit restores the
behaviour of continuing where reasonable to do so but at the same time
introduces a log message so there is a record that the file open
failed.