mem: Issue #12 - enable ROMS to be loaded in a defined order.
This causes two passes to be made over the ROM files in the folder for
a given model.  The first pass picks up all ROMs whose filename starts
with a number.  Assuming the number is in-range the slot is still free
the ROM is loaded into the specified slot.  A second pass loads any
remaining ROM files into the remaining vacant slots in no particular
order.