vdfs: Stop the ROM claiming OSWORD 7F when not the current filing system
The VDFS ROM includes code to respond to OSWORD 7F and leave a memory
signature in memory ordinarily used by DFS to avoid certain programs
from detecting that they are not running under DFS.  Unfortunately it
did not check that it was the currently selected filing system so in
cases where VDFS is available but the user has chosen to boot into
DFS, VDFS was claiming DFS's OSWORD calls.  This commit fixes that.