65816: Fix failure to remember 65816 is enabled.
There is a test for the OS ROM needed by the 65816 and if the OS
ROM is not found the 65816 is disabled again.  This test was
written with a pathname that uses Windows path separators which
Linux does not understand.  This commit changes them to Unix-style
separators because the Win32 does understand those.