debugging: re-introduce traps including 6502 BRK
Prior to the work to implement a non-CPU-specific interface between
the debugger and the CPU modules the 6502 implemented a trap to the
debugger when the BRK instruction was hit.  The commit re-introduces
that but using a new feature of the non-CPU-specific interface that
enables any CPU to raise one of a number of CPU-specific traps.  This
could be extended to IRQ and NMI and also to similar events on other
processors.