debugger: variable size read/write/input/output break/watchpoints
This change causes the debugger to use a different number of hex
digits for the values printed by read, write, input and output
watchpoints and breakpoints in response to the information from the
CPU on the width of memory access being performed.  Previously it
would always have used sufficient digits but a 16 bit access for the
value of zero would have been printed as 00, now it will be 0000.