debugger: add a -debugtube option in place of the CPU menu.
The CPU menu appeared starting the debugger and allowed any CPU to be
enabled for debugging but if a CPU was chosen that was not running no
debugging would happen.  A new -debugtube option automatically enables
debugging on the current tube processor instead with -debug reverting
to debugging the core 6502 processor.