debugging: fix issues with extra video refresh and single-stepping.
oss003 commented in http://stardot.org.uk/forums/viewtopic.php?f=4&t=10823&start=240#p165842
that the extra video refresh was not working in single-step mode.  I
found that the extra refresh was performed but as this had the effect
of generating an interrupt the next instruction after any single-step
was the start of the interrupt handler.  This commit adds a parameter
to video_poll so it can be run without generating an interrupt and
that is exactly what the debugger then does.