debugger: Make writem write to right CPU memory and add extra vrefresh
This commit fixes a bug with the writem debugger command which would
always write to the core 6502 processor memory even when debugging a
tube processor.  At the same time is also adds the extra video
refresh, if the processor is the core 6502, so if the write was to
screen memory the effect can be seen.