discs: namespace cleanup
The functions that operate on the various disc formats are assigned to
a structure when the media of a particular format is loaded by the
load function for that format.  These functions do not, therefore,
need external linkage because they are access only indirectly via that
structure.  This commit makes them static instead.