wd1770: implement verify on type I command for WDDFS
The Watford DDFS issues the restore command with the verify bit set
and it is during the execution of the verify phase of the restore
command that it expects to receive a sector not found error if the
density of media does not match the density select it has selected.
Delaying the sector not found error to the read sector operation does
not work.  This commit implements the verify option on restore and the
other type I commands, i.e. those that move the disc head.