models: check model number before array access.
This bug appeared when using a config file from a later branch of B-em
with more models on an earlier branch with the result that B-Em simply
crashed due to an array bounds error.  This commit adds code to check
the model number after reading from the config file or command line
and forces out of range values to sane ones.