tube: resolve issue#30: 256 byte transfer over Tube hangs
See https://github.com/stardot/b-em/issues

The issue was caused because the original implementation of the tube
ULA emulation in B-Em assumed the status bits work identically for all
the registers, consistent with documention on mdfs.net, but this causes
the code in the tube client ROM to wait forever before beginning a 256
byte transfer.  It turns out register is odd in the real hardware so
this change is to better emulate the real hardware,  brings B-Em into
line with the way PiTubeDirect emulates the tube and fixes the issue
with the 256 byte transfers.