vdfs: match file numbers used by vdfs.c vs. vdfs.s
The VDFS ROM responds to service ROM call 25 with a set of file system
information blocks for the Master file system switch but the one for
VDFS was using a different range of open file numbers than the C code
in vdfs.c that actually handles opening the files.  This change brings
them into line.