vdfs: connect SRAM parsing and execution internally when VDFS is current FS
Previously VDFS would parse the SRAM commands SRLOAD and SRWRITE into a
standard OSWORD &43 paramater block whether or not it was the current
filesystem and would return to the VDFS ROM which would then decide
whether to issue a standard OSWORD call for execution, if the current
filesystem is not VDFS, or return to the VDFS host module for execution
if it is.

The new approach is that the VDFS host module checks if it is the
current FS and, after parsing the command to internal variables,
continues to execution if it is the current FS and then returns with
carry clear.  If it is not the current FS it writes the variables into
the OSWORD parameter block and returns carry set so the VDFS ROM knows
to issue the OSWORD call.