m4000: Separate and improve Linux MIDI implementation.
This moves the Linux MIDI implementation into a separate file from the
Music 4000 keyboard emulation and adds an ALSA raw MIDI implementation
to compiliment the existing JACK and ALSA sequencer options.  It also
uses conditional compilation to compile the implementations for which
the headers are present on the system being compiled and allows run-time
selection of MIDI implementations via the b-em.cfg file.