m4000: make Music 5000 buffer size, and therefore latency, configurable.
This restores the default buffer size for the Music 5000 to 3000 which
is good for avoid the sound breaking up on machines with limited CPU
power.  It then adds a config file option in the [sound]
section with theyt key 'buflen_music5000' which can be used to tune the
buffer for lower latency by reducing its size for those interested in
the M4000 emulation.
[sound] section