Re-work the way files are loaded and saved to make Linux packaging easier.
This looks for data files, including ROMs and noise samples one set of
directories, so these may be provided as part of an installed package
and then augmented by the user, and config files in another set of
directories so these may be user-specific.

This also enables the set of ROMs loaded for a particular machine to be
overridden by the config file, rather than being determined by the set
of ROM files in a particular directory.