vdfs: improved handling of disappearing host directory.
This deals more effectively with the case where the current Acorn
directory corresponds to a host directory that has been deleted or
otherwise become unreadable.

Firstly it avoids an infinite loop between the VDFS ROM and the VDFS
ROM that previously occured when trying to catalogue the current
directory in that situation.  Now it generates a "Not found" error
instead.

Secondly it changes the way file/directory names are parsed so it is
possbile to recover from this situation by changing directory from
within the BBC guest either to the parent of the deleted directory or to
the root.