m2000: add config variables and Linux GUI options for Linux MIDI for M2000
This creates some configuration variables for each of the Music 2000
outputs and each of the Linix MIDI sub-systems, arranges these config
variables to be loaded from and saved to the config file and also
provides the Linux GUI to choose these.  It also renames the config
variables from the Music 4000 for consistent namimg.