vdfs: make *LIST and *TYPE cope with non-Acorn line endings.
This copies a feature of the Master MOS version of these two command in
they can cope with files whose lines are delimited with either CR alone,
as per Acorn and MacIntosh, LF alone, as used on Unix, and CR+LF as used
on DOS and Windows.

It also introduces the *PRINT command which simply copies a file to
OSWRCH without any translation or any notion of lines.