vdfs: more thorough overhaul of catalogue size counting.
The complexity here is in dealing with a file entry that is being marked
as no longer extant and therefore does not need to be included in the
count of files for cataloguing and the case where a file entry is in the
tree but marked non-existant and then needs to be marked extant again.

This version introduces two new functions one to be called when a file
or directory is created and one when a file or directory is deleted that
do the necessary accounting taking into account the previous state of
the entry concerned.