vdfs: fix OSFILE saving file with full name in the wrong place
When writing a file with OSFILE, if the file did not previously exist it
was being created in the current directory even if the filename
contained the name of the directory in which it should be created.  This
is a regression as it used to work correctly.  This commit fixes it.