6502: use correct flag to set CPU to disassemble.
Prior to this change the 6502 module called the 6502 disassembler with a
flag indicating whether the CPU was a 6502 or 65C02 based on whether the
machine was a master or not.  This causes an issue if a configuration is
created that has a 65C02 in an otherwise non-master machine.

This commit changes the 6502 module to use a separate global x65c02
which is set by the models module.