Fix autoboot holding the shift key down too long.
Lurkio produced a test SSD which enabled autobooted a BASIC program
which then engaged paged mode (VDU 14) which appeared to be ignored
because the simulated shift key press for autoboot was persisting long
after the boot had happenned.

This commit resolves the by cancelling the simulated keypress as soon as
the filing system accesses the disk or, in the case of VDFS, once it has
started to initialise beyond the point it has scanned the keyboard.