allegro5: work around broken al_get_path_extension function.
The Allegro 5 al_get_path_extension returns the extension including the
initial '.' which is not what one would expect from a function that is
supposed to be OS independent.  This fix skips past any initial '.' on
the extension so that disc/tape filenames once again match the table of
loaders.