sdf: add DOS file formats by searching for format by size as last resort.
Most disc formats are reconised by finding bytes within the file system
layout on the disc that match a table of geometries in this module.
There are some DOS and Hybrid ADFS+DOS formats where this is not
possible which were recognised by size in disc.c prior to the
introduction of SDF.  This commit re-instates that, via table, but only
when no other format has matched.