allergo5: Fix Issue#51
Because Allegro 5 can work with multiple config files, when B-Em does
not find a config file the config file functions to not just work with
an empty configuration.

B-Em creates an empty config to save values into but only at config save
time as part of shutdown and only when it has found somewhere suitable
to save the file.

Taken together these mean that config values should not be written
during normal running, only at config save time.  This commit aims to do
that.