allegro5: reduce CPU used by video blitting.
When first porting the video output to Allegro 5 using a bitmap which
automatically converted into a video bitmap and then manipulating it via
the al_get_pixel/al_put_pixel had proved to be too slow and a speed
improvment had been obtained by using a memory bitmap.

Then switching to locking the bitmap and manipulating it directly with
functions which the compiler would inline made a further big improvement.

Now it turns out with the bitmap locked the speed penalty for drawing
into a video bitmap disappears and a video bitmap is faster blitting to
the backbuffer, thus going back to using a video bitmap reduces the CPU
used in video emulation and enable a faster potential emulation speed.