allegro5: re-introduce tube speed option and address Issue #54 3Mhz/4Mhz
In Issue #54 Hoglet notes that the external 65C02 co-processor runs at
3Mhz wheras the internal 65C102 in the master runs at 4Mhz.  Because the
tube speed code was tied to powers of two the extern processor was being
run at 4Mhz equivalent.  This commit changes tube_shift into a multipler
which, coupled with a right shift (divide by two) enables the speed of a
tube processor to be other than a power of two multiple of the core 6502
speed.  Accordingly the speed of the 6502 tubes are now correct.

This commit also re-introduces the GUI option for changing the tube
speed multiplier but this time it is in forcce of any tube processor,
not just the 6502.