allegro5: new framework for save/load state including tube.
This new framework for saving and re-loading state records a key letter
and a length in the file for each section saved.  This means if one
section does not reload the same number of bytes as it previously saved,
for example because of an enhancement to the emulation of the hardware
concerned, it does not throw the whole file out of sync.

This also includes the framework to save and load the state of tube
processors although this commit does not implement this for any of the
processors concerned, just for the tube ULA.