allgro5: fix the speed scaling (of non-tube BBC) by fixing the frame skips
Blitting a screen to the host display introduces a wait for frame sync
delay which limits the speed at which the emulated machine can run.  In
order to go faster than 100% it is necessary to drop frames and
vid_fskipmax facilitates that but strangely 1 means "dont skip any
frames" so the values in the table of speeds were off by one.