paste: change from OSRDCH to intercepting keyboard buffer remove.
This has the core 6502 learn the address of the routines to remove a
character from a buffer and count the contents of a buffer and then
intercepts these when there is text from the clipboard to paste.

Compared to the previosu implementation this has the advantage that
pasting auto-starts.  Intercepting OSRDCH required a key to be pressed
to make a fresh call to OSRDCH before pasting would start.