savestate: save/load complete model description to state file.
For the version 2 snapshot file this saves both the index of the current
model within the table of models in the copy of B-Em that saved the
snapshot and the complete description of the model concerned.

On loading a snapshot the description is matched to the models in the
table of models in the loading B-Em and if not found a new model is
added specifically to run the snapshot.

This enables a user to define a personal model and then have it work
with saved snapshots.