keyboard: enable keys interpreted by B-Em to be redefined.
This changes to feeding a key event through the user-defined key mapping
before testing for the keys that B-Em itself intercepts rather than
afterwards.  That means that the Break key can be re-assigned, and the
keyboard mapper GUI is enhanced to do that, but also via editing the
config file any other B-Em key can be re-mapped.