joystick: Add support for mapping joystick sticks, axes and buttons
This enabled the various axes on the various sticks within a joystick
device to be mapped to one of the BBC ADC channels and have an
associated scaling factor.  Joystick axes can also be mapped to key so
moving the stick more than halfway in that direction simulated a
keypress.  Finally, buttons may be mapped to one of the fire buttons as
seen by the BBC or likewise mapped to keypresses.