sdf: Fix Issue#63.
The issue was that the simple heuristic of checking for DFS catalogue on
a potential second side with the same number of sectors was getting a
false positive because the byte concerned happen to match even though
there was no 2nd catalogue.

This commit adds further checks, i.e. the the number of directory
entries declared in each catalogue is sane and that the files in the
catalogue are sorted in descending order of sector start address.  This
makes it much less likely that data sector will be mistaken for a
catalogue.