sdf: Fix loading of sequential sides media.
This bug was a regression caused by adding a file size check in commit
ad6549800f8d8013f6af62b6cc6ea95ba1b1ad76 to fix the loading of DFS+CP/M
hybrid discs.  Unfortunately the size used in the check was always the
size for a single-sided disc.