video: Fix Issue#65.
This was caused by MOS 3.50 programming values into the CRTC which
caused the X direction to go as far as 1280, the width of the target
bitmap but because mode7_render also used 16 further pixels to the
right of this, and because there are no bounds checks in put_pixel, this
resulting in writing over memory outside the bitmap.

The fix is to constrain the X direction to 16 less than 1280.