vdfs: implement OSWORD 7F but fallback on disc image files.
Previously, when OSWORD 7F was issued when VDFS was the current filing
system VDFS did not claim the OSWORD ROM service call in the hope some
other filing system would implement it.

Now VDFS implements the call but, as the concept of raw sector
read/write doesn't map neatly to host files this uses the disc images
loaded for FDC emulation with the exception that only the SDF module
is supported.  The SDF module handles "simple" formats such as SSD/DSD.