Issue#69: Fix "Eject Disc".
Prior to this commit Eject Disc was closing the disc so to a program
running on the emulated machine it would appear the drive was empty but
was retaining the name of the file.  Next time the emulator was
re-started the previous disc image would be re-opened.

This change discards the name of the disc image and fixes the config
file handling so it is not re-established when B-Em is re-started.