Issue#71: split DD formats from SSD/DSD, use extension and size.
Prior to this commit, in order to cope with truncated images, the SDF
module would look into a disc image file and try to determine both the
type of filesystem and the size.  It would then look these up in a
series of tables to work out the likely disc geometry.

Unfortunately, there are some disc images in the wild with non-standard
filesystem sizes which would cause SDF to complain about not being able
to determine the geometry.

Part of the reason for the previous approach was to spot cases where the
contents of the disc image were from a double-density (MFM) disc - this
would be apparent because of the size of the filesystem.

This commit returns to treating SSD/DSD files as single density only,
i.e. having 10, 256 byte sectors per track.  The only heuristics that
remain are checking for a double-sided sequential sides disc in an SSD
file (if the size is greater than 80 tracks) and a DSD file being
sequential sided rather than the more usual interleaved sides.  These
checks do rely on being able to detect the DFS catalogue sectors.

Non-ADFS double-density discs, i.e. those used by Opus/Solidisc/Watford
DDFSes are now expected to be in .sdd and .ddd files which are not
permitted to be truncated to the size of the host file can be used to
determine the format.  This convention is copied from b2.

See these discussions on StarDot:
https://stardot.org.uk/forums/viewtopic.php?f=4&t=15651
https://stardot.org.uk/forums/viewtopic.php?f=4&t=13081&start=90#p214039