sdf: split access and geometry code and provide test harness
This splits the previous sdf.c module into two.  One, sdf-acc.c provides
the access functions including state machine and the tother, sdf-geo.c
provides geometry recognition and the creating of new disc images.

This enables the sdf-geo.c module to be linked into a test harness to
check the recognition is correct for a whole set of discs independent of
the main B-Em executable.