vdfs: fix crash when trying to create file in non-existent dir
See discussion at
https://stardot.org.uk/forums/viewtopic.php?f=4&t=15834&p=217704#p217690

The crash was caused by failure to handle the case of trying to create a
file in a non-existent directory.  In vdfs_write ent is passed in and is
the result of seacrching for key, NULL in the case that the file does
not exist.  A previous call to find_entry has set key->parent to the
directory in which to create a new file or NULL in the pathname
specifies a directory that does not exist.  osfile_write did not take
account of this and passed this NULL to add_file which attempted to
dereference it.