vdfs: Fix bug in VDFS-specific read-filenames OSGBPB.
The convention between the VDFS ROM and the vdfs.c module for listing
filenames in a directory is a little odd and doesn't match the same
OSGBPB function as standardised by Acorn.  Attempting to apply the same
change to both resulted in the VDFS *CAT command giving an output that
is one entry short.  This fixed that.