vdfs: incomplete, make vdfs.c more master, and vdfs.asm more slave.
This moves some of the logic that was previous in the VDFS ROM as 6502
assembler into the vdfs.c module making the vdfs.c module more of a
master and the VDFS ROM more of a slave.  In particular the vdfs.c
module now sees all ROM service calls.  This adds the potential for
further integration in the future.