Add Master models ROMs patched for Y2K and OSW7F
This commit includes patched versions of both MOS 3.20 (MOS 3.22) and
MOS 3.50 (MOS 3.52) and model definitions that use these ROMs instead
of the vanilla ones.  In each case a Y2K patch is included so the real
time clock works in the 21st century and in each case DFS is patched so
it no longer reponds to OSWORD 7F (floppy disc access) if it is not the
current filing system.

Without the DFS patch, DFS responds to OSWORD 7F but then makes use of
workspace it has not claimed because it is not the current filing
system.  This means both that the OSWORD call can fail to work properly,
and I found this try to boot CP/M on the Z80, and also corrupt memory it
does not own.

With the patch, DFS ignores OSWORD 7F.  On real hardware this can be an
issue if other code relies on sector level access to floppies.  That
includes the Z80 and 80186 boot code and also HADFS, but on B-Em VDFS
also supports OSWORD 7F, and uses the same disc image files as would be
accessed by DFS, but without relying on any workspace in the emulated
BBC.