vdfs: *DISC/*ADFS not selecting their proper filing systems.
The bug here was that even after having issued a:

*FSCLAIM OFF

command and confirming that setting had changed, issuing a *DISC or
*ADFS command had no effect and VDFS remained as the select filing
system.

This was caused by always returing from ROM service call 04,
unrecognised command, with A=0 even if the VDFS code opted not to
do anything in response to that command.  That in turn avoided any other
ROM examining the command so DFS/ADFS simply did not see it.

This commit enables vdfs_do to return whether it has handled the action
(true) or whether it should be passed on (false) and uses that to
control the value of A.