vdfs: replace binary trees for directory contents with a linked list.
This replaces the pair of binary trees used to hold the contents of a
directory so that it may be quickly searched in order by either Acorn
filename or host filename with a linked list.

This will not be quite so quick to search but should use less RAM and be
an easier starting point to implement wild card matching.