via: always clear T1 interrupt when T1LH is written.
Previously, when the T1LH was written, the T1 interrupt was only cleared
if the timer was in continuous mode.  It turns out a real VIA is not
that complex and always clears the T1 interrupt when T1LH is written.

See https://stardot.org.uk/forums/viewtopic.php?p=223734