via: minor changes, inline and TLMIT constant.
Prior to this change the function via_poll was an in-line in the header
via.h such that it could be inlined into polltime in 6502.c.  Running
under gdb has shown this did not normally save a function call so it
made more sense to make via_poll a function with external linkage in
via.c where it could possibly have via_updatetimes inlined into it.

The TLIMIT constant was also introduced for the value of the VIA
counters at which they are considered to have expired and set to -3,
the value that was previously hardcoded in various places.