Use key_paste_start() approach for logical keyboard
The code has been quite heavily reworked but I think the influence of
key_paste_start() is clear. The main difference is that for logical
keyboard mode, we need to take account of the host's SHIFT/CTRL keys *at
the time the key is pressed*; they may well be different by the time we
come to type that character on the emulated machine's keyboard. (When
pasting from the clipboard this isn't an issue; the clipboard contents
completely determine which emulated keys should be pressed.)