video: re-work display modes.
This introduces a new display mode "scaled" which is similar to line
doubling but instead of B-Em duplicating the lines it asks Allegro to
scale up the bitmap in the vertical direction.  Some graphics systems
may do some interpolation and it may be more efficient than copying
rows.

This also splits the user chosen display mode from the one used
internally.  This allows the mode used internally to be forced to
something sensible when the user chosen mode is "Interlaced" but
interlace is turned off in CRTC register 8 (e.g. via *TV0,1).  In that
case the display mode used internally is forced to "Scaled".

Finally, this extends PAL emulation to all the display modes.