video: fix junk on screen after border change, Issue#86.
See https://github.com/stardot/b-em/issues/86 for discussion.

When changing border size from medium to none with MODE 7 (teletext) in
effect, some fragments of the screen appeared to be copied to the very
bottom of the window.  This is because even in "border none" there is a
very small border.  This in turn is due to the window size being chosen
to be able to fit a graphics rather than teletext screen for the style
of border concerned.  The border concerned was not being drawn to so
retained its previous contents which was a part of the displayed screen.

This commit fixes the issue by filling this border with the border
colour (preset to be black).