vdfs: Fix "No filing system" on MOS 3.20
On the Master with MOS 3.20, in the tail of the ROM service call code,
if the service call was &0F (filing system vectors changed) it issues an
OSARGS call to get the new filing system number and caches this at
&DF00.  It then replaces the OSARGS handler with a filing system switch
one which returns the cached value rather than letting the filing system
handle it.

The VDFS OSARGS handler, implemented in vdfs.c, returns the value of
fs_num but prior to this commit the OSFSC call to warn of impending
filing system change (A=&06) was clearing this value.  This commit has
the assembler routine in the ROM set the right fs_num after it has
issued this OSFSC call and installed the correct vectors but before
issuing the ROM service call to say vectors have changed.