tube: refactor tube initialisation.
This moves a number of tube processor specific functions out of tube.c
which is mainly concerned with the Tube ULA into the files that
implement the processor concerned.  This also gives the opportunity to
expose fewer functions from each CPU implementation.

Also, the tube initialisation in model.c will load the ROM from the disk
file and pass it to the processor initialisation function.