RTC: add save/restore of Epoch.
When the user does not explicitly set the date/time B-Em simply reports
the host system clock.  If the user does set the date/time, instead of
recording the date/time concerned B-Em adjusts the Epoch to which the
system time is added so the set date/time then increments with the host
system clock.

This commit saves and restores that Epoch so when B-Em is re-started it
is as if the set date/time was actually saved and then kept good time in
the meantime.