pdp11: avoid emulator hang on PDP11 panic.
Previously, when the PDP11 core detected an error and called its own
panic() function this would spin in a do-nothing loop which would hang
the whole emulator by never returning and therefore never allowing any
other part of the emulator to run.

This change allows the PDP11 core to return from the panic but then
heeds its own halted flag so the in any subsequent call the PDP11 core
does not attempt to do anything.