tube: Fix issue#91 by returning 1 for unused bits of R3 status.
See the issue on GitHub for more detail but the way the parasite to host
transfer type 6 is coded in Pandora relies on the unused bits of the
tube R3 status as seen from the parasite (pstat[2]) being 1 rather than
0.  When those bits are set to 0 a test for R3 being ready to write to
fails to detect R3 being full and the parasite keeps writing to the
already full R3 before the host has a chance to read it.

Transfer type 6 is used by ADFS so anything written to ADFS would be
corruped and this would also crash the 32016 when installing Panos from
floppies.