integra: fix failure to enter reset code on Ctrl-@-Break.
See https://stardot.org.uk/forums/viewtopic.php?f=4&t=17675&p=249181#p248995

There was an issue where Ctrl-@-Break would offer the option to reset
the configuration on real hardware but not in B-Em or in BeebEm.  This
turned out to be due to the failure to emulate the resetting of RTC
registers B and C.  This commit adds an implementation of this.