6502: emulate partial address decoding of ROMSEL on the BBC B
See https://stardot.org.uk/forums/viewtopic.php?f=4&t=17839

On the BBC there is only partial address decoding so the whole of the
range &FE30 to &FE3F writes to ROMSEL.  B-Em previously had the decoding
as per the Master where this range is decoded into blocks of four
addresses.  This commit make B-Em change the way this area is decoded
according to which model it is emulating.