disclock: add linux-only file locking for SDF disc images.
This is implemented on Linux only at the moment as on Windows files
opened by fopen are not shared by default.

This change causes the disc image file to have an advisory, exclusive
(write) lock placed on the whole file when the FDC spins up the disc.
When the disc spins down again the lock is released.  This will enable
two or more copies of B-Em with the same disc or discs mounted to see
each others changes and avoid corruption.