Issue#94: Fix core dump loading a non-config-file model.
When a snapshot file loaded if the model saved in the snapshot file is
not one of the models in the list loaded from the b-em condig file the
model from the snapshot file is added to the list of known models.
Because this new model had not come from the config file its cfgsect
element was NULL which caused the core dump when subsequently
dereferenced.  This commit allocates a new config section name for the
newly loaded model.