Fix Issue#96: Write protect tickbox out of step.
The tickbox in the Disc menu was getting out of step with the internal
flag that is actually used by the disc system to enforce a read-only
disc.

This happenned in two places firstly, as per the bug report, when
creating a new disc.  In this case the bug was that the menu being
searched for the appropriate read-only item was the sub-menu for the
disc type rather than the top level disc menu hence the menu item was
not found and the flag not set.

The other issue was that the menu tickbox was not being set when the
read-only status of the discs was being set at startup time after the
names of the discs to load have been taken from the config file or on
the command line.

Both are fixed by the GUI providing a new function to set the read-only
status and calling it from the other two places.