Issue#96: Set write protect ticks correctly on start-up with r/o image.
When a disc image file is read-only, i.e. when B-Em fails to open it
read/write but suceeds in opening it read-only it is treated as a write
protected disc, i.e. as if the write-protect notch was taped.  When this
occured during emulator initialisation the read-only tickets on the disc
menu were not being set.  This commit fixes that.