Fix compilation on GCC v10.
On prior versions of GCC a non-static variable could be defined more
than once and all instances would be coallesced.  This no longer applies
in GGC 10, at least with the default options.  Thus this commit deals
with all the variables for which GCC complains there are multiple
defintions by either making sure the declaration in the header file is
not also a definition by adding 'extern' and then if necessary adding a
single definition within one of the C files, or my making the variable
static within each C file in which it occurs when the variable does not
need to be program-global.