debugger: ROM-specific breakpoints and address formatting.
This commit introduces ROM-specific breakpoints (and watchpoints) for
the core (I/O) processor.  This works by having all breakpoints as a
minimum of a 32 bit address and having the 6502.c module fold the ROM
number into the address before passing to the debugger.  Thus the least
signficant 16 bits of the address passed to the debugger are the actual
6502 address and bits 17-20 are the ROM number.

To make this a little easier to use this commit also introduces
processor-specific address formatting by adding a method to the
cpu_debug_t structure.  Some generic address printing functions are
provided in debugger.c and most CPUs use those, but odd cases can use an
internal function.  6502.c uses a custom function to print the ROM
number and x86.c uses a custom function to print segmented addresses.