imd: implement write track for Acorn 1770 DFS
This adds support for the WD1770 write track to IMD.  The WD1770
receives all the track data from the host rather than generating it and
we parse it here to extract the necessary ID values to set up
tracks/sectors in memory ready to be written out to the file.

This works fine with Acorn 1770 DDS but may not yet work on others.