vdfs: Close SPOOL/EXEC files on OSFIND A=0,Y=0
When OSFIND is called with A=0 (close file) and Y=0 (all files) it is
possible one of the files it is being asked to close is a current SPOOL
or EXEC file.  With EXEC files, at least, this can result in an infinite
loop of "Channel" errors.  This commit has VDFS issue an OSBYTE &77 to
close any SPOOL/EXEC files before actually closing all open files.

Although this is not perfect given that files could be open on multiple
filing systems it seems to be the accepted practice.  See the discussion
at https://stardot.org.uk/forums/viewtopic.php?f=54&t=19949