leds: initialise LEDs along with everything else.
Prior to this commit the LED bitmap was allocated and drawn into on the
first screen blit called from the video emulation because doing this
initialisation from main_init as most modules work resulted in the drawn
line/background not appearing on screen.  The problem turned out to be a
al_clear_to_color in video_poll.c that was not setting the current
bitmap.  Prior to LEDs this happenned to be right after the call to
video_init but LEDs left it at led_bitmap so the video code then cleared
the newly drawn LED bitmap.  This fixed the bug in the video emulation
and then moves LED initialisation to its proper place.